+++ bolt/lib/nonlinear/finite_volume/fvm_operator.py
@@ -14,7 +14,7 @@ def op_fvm(self, dt):
     Parameters
     ----------
 
-    dt : float
+    dt : double
          Time-step size to evolve the system
     """
